requirements.txt
streamlit==0.77.0
pandas~=1.2.2
numpy~=1.20.1
pillow~=8.1.0
scipy~=1.6.1
